feat(formulas): Math module defined
