+++ src/lib/ui/app/san-formulas/math/index.ts
@@ -1,0 +1,13 @@
+import { math } from './core.js'
+import { MathOperators } from './_operators.js'
+import { MathIndicators } from './_indicators.js'
+
+math.import(
+  {
+    ...MathOperators,
+    ...MathIndicators,
+  },
+  { override: true },
+)
+
+export { math as SanFormulas, Timeseries } from './core.js'
